# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CRM Cia da Vacina API
  version: 0.1.0
  description: |
    Contrato Sprint 1+ — Auth, Users, Units, Me.
    Endpoints de Inbox/WhatsApp entram nas sprints seguintes mantendo este estilo.
servers:
  - url: http://localhost:8080/api/v1
    description: Local API
...
paths:
...
  /inbox:
    get:
      tags: [Inbox]
      summary: Lista conversas do inbox (contrato antecipado p/ FE mocks)
      operationId: listInbox
      security:
        - bearerAuth: []
      parameters:
        - name: unit_id
          in: query
          schema:
            type: string
            format: uuid
        - name: stage
          in: query
          schema:
            $ref: '#/components/schemas/PipelineStage'
        - name: owner
          in: query
          description: me | unassigned | UUID
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/InboxItem'
                  next_cursor:
                    type: string
                    nullable: true
  /conversations/{conversationId}/claim:
    post:
      tags: [Conversations]
      summary: Atendente assume a conversa (handoff)
      operationId: claimConversation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: Claimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboxItem'
        '409':
          description: Já atribuída a outro
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/pipeline:
    patch:
      tags: [Conversations]
      summary: Atualiza etapa do pipeline
      operationId: updatePipeline
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stage]
              properties:
                stage:
                  $ref: '#/components/schemas/PipelineStage'
                reason_code:
                  type: string
                reason_text:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboxItem'
        '422':
          description: Motivo obrigatório para nao_fechado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: unauthorized
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    PipelineStage:
      type: string
      enum:
        - em_atendimento
        - em_negociacao
        - aguardando_fechamento
        - fechado
        - nao_fechado
...
    InboxItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contact_name:
          type: string
        contact_phone:
          type: string
        unit_id:
          type: string
          format: uuid
        pipeline_stage:
          $ref: '#/components/schemas/PipelineStage'
        mode:
          type: string
          enum: [ai_triage, human]
        owner_id:
          type: string
          format: uuid
          nullable: true
        intent:
          type: string
          nullable: true
        ai_summary:
          type: string
          nullable: true
        last_message_preview:
          type: string
        last_message_at:
          type: string
          format: date-time